refactor(chat): aws_bedrock 拆分 payload/stream 子模块
将 payload/aws_bedrock.rs（~1.4K 行）与 infrastructure
/http_chat_completion_repository/aws_bedrock.rs（~1.8K 行）拆成
provider 子模块：

- payload/aws_bedrock/{anthropic,nova,llama,mistral,deepseek,cohere,
  ai21_jamba,shared}.rs：每个文件负责一个 provider 的 invoke body
  构造与 helper；shared 集中 OpenAI messages → flat-message 工具
  与 passthrough chat-messages 助手
- infrastructure/.../aws_bedrock/{nova,llama,mistral,deepseek,cohere,
  ai21_jamba}.rs：每个文件负责该 provider 的非流式响应 Claude-shape
  投影与 EventStream chunk → Anthropic content_block_delta 转换

根模块只保留 dispatch、Provider 枚举、Bedrock invoke/stream endpoint
处理、catalog 合并、EventStream 解码骨架与跨模块集成测试，单文件
体积控制在 ~350 / ~950 行。功能完全等价，全部测试用例随实现一起
迁移到对应子模块，主文件保留 dispatch / catalog 集成测试。
