+++ src-tauri/src/infrastructure/apis/http_chat_completion_repository/aws_bedrock.rs
@@ -1,8 +1,22 @@
+//! AWS Bedrock infrastructure adapter.
+//!
+//! Orchestrates HTTP calls to Bedrock's `/model/{id}/invoke` and
+//! `/invoke-with-response-stream` endpoints plus the regional control plane
+//! catalog endpoints (`/foundation-models`, `/inference-profiles`).
+//!
+//! Each provider's request body construction lives in
+//! `application::services::chat_completion_service::payload::aws_bedrock::*`;
+//! their response / stream-chunk projections live in the sibling submodules
+//! (`anthropic`, `nova`, `llama`, `mistral`, `deepseek`, `cohere`,
+//! `ai21_jamba`). Everything here is the routing glue: pick the right
+//! provider, run the HTTP request, drain the EventStream, and forward
+//! normalized text frames upstream.
+
 use base64::Engine;
 use base64::engine::general_purpose::STANDARD as BASE64_STANDARD;
 use reqwest::RequestBuilder;
 use reqwest::header::{ACCEPT, CONTENT_TYPE};
-use serde_json::{Map, Value, json};
+use serde_json::{Value, json};
 
 use crate::domain::errors::DomainError;
 use crate::domain::repositories::chat_completion_repository::{
@@ -13,6 +27,13 @@ use crate::domain::repositories::chat_completion_repository::{
 use super::HttpChatCompletionRepository;
 use super::normalizers;
 use super::response_body::read_upstream_json_body;
+
+mod ai21_jamba;
+mod cohere;
+mod deepseek;
+mod llama;
+mod mistral;
+mod nova;
 
 const BEDROCK_PROVIDER_NAME: &str = "AWS Bedrock";
 const BEDROCK_EVENTSTREAM_CONTENT_TYPE: &str = "application/vnd.amazon.eventstream";
@@ -288,88 +309,28 @@ fn normalize_provider_response(
     match detect_http_provider(model_id) {
         HttpBedrockProvider::Anthropic => normalizers::normalize_claude_response(body),
         HttpBedrockProvider::Nova => {
-            normalizers::normalize_claude_response(nova_response_to_claude_shape(body))
+            normalizers::normalize_claude_response(nova::response_to_claude_shape(body))
         }
         HttpBedrockProvider::Llama => {
-            normalizers::normalize_claude_response(llama_response_to_claude_shape(body))
+            normalizers::normalize_claude_response(llama::response_to_claude_shape(body))
         }
         HttpBedrockProvider::Mistral => {
-            normalizers::normalize_claude_response(mistral_response_to_claude_shape(body))
+            normalizers::normalize_claude_response(mistral::response_to_claude_shape(body))
         }
         HttpBedrockProvider::DeepSeek => {
-            normalizers::normalize_claude_response(deepseek_response_to_claude_shape(body))
+            normalizers::normalize_claude_response(deepseek::response_to_claude_shape(body))
         }
         HttpBedrockProvider::Cohere => {
-            normalizers::normalize_claude_response(cohere_response_to_claude_shape(body))
+            normalizers::normalize_claude_response(cohere::response_to_claude_shape(body))
         }
         HttpBedrockProvider::Ai21Jamba => {
-            normalizers::normalize_claude_response(ai21_jamba_response_to_claude_shape(body))
+            normalizers::normalize_claude_response(ai21_jamba::response_to_claude_shape(body))
         }
         // Other providers reuse the Claude normalizer for now (they fail at the
         // application layer before reaching this point); kept for forward
         // compatibility with custom-template overrides.
         HttpBedrockProvider::Other => normalizers::normalize_claude_response(body),
     }
-}
-
-/// Reshape an Amazon Nova non-stream `invoke` response into the Claude-style
-/// `{ content: [...], stop_reason, usage }` envelope that
-/// `normalize_claude_response` already understands. Doing the translation here
-/// (rather than building a parallel normalizer) keeps the OpenAI-shape choice
-/// the frontend sees identical to the Claude path.
-///
-/// Nova's response shape (Converse-style even when called via `/invoke`):
-/// ```json
-/// {
-///   "output": { "message": { "role": "assistant", "content": [{ "text": "..." }] } },
-///   "stopReason": "end_turn",
-///   "usage": { "inputTokens": N, "outputTokens": M, "totalTokens": N+M }
-/// }
-/// ```
-fn nova_response_to_claude_shape(body: Value) -> Value {
-    let mut content_blocks: Vec<Value> = Vec::new();
-    if let Some(parts) = body
-        .pointer("/output/message/content")
-        .and_then(Value::as_array)
-    {
-        for part in parts {
-            if let Some(text) = part.get("text").and_then(Value::as_str) {
-                content_blocks.push(json!({ "type": "text", "text": text }));
-            }
-        }
-    }
-
-    let stop_reason = body
-        .get("stopReason")
-        .and_then(Value::as_str)
-        .unwrap_or("end_turn")
-        .to_string();
-
-    let mut usage = Map::new();
-    if let Some(usage_value) = body.get("usage").and_then(Value::as_object) {
-        if let Some(input_tokens) = usage_value
-            .get("inputTokens")
-            .or_else(|| usage_value.get("input_tokens"))
-            .and_then(Value::as_u64)
-        {
-            usage.insert("input_tokens".to_string(), json!(input_tokens));
-        }
-        if let Some(output_tokens) = usage_value
-            .get("outputTokens")
-            .or_else(|| usage_value.get("output_tokens"))
-            .and_then(Value::as_u64)
-        {
-            usage.insert("output_tokens".to_string(), json!(output_tokens));
-        }
-    }
-
-    let mut claude_body = Map::new();
-    claude_body.insert("content".to_string(), Value::Array(content_blocks));
-    claude_body.insert("stop_reason".to_string(), Value::String(stop_reason));
-    if !usage.is_empty() {
-        claude_body.insert("usage".to_string(), Value::Object(usage));
-    }
-    Value::Object(claude_body)
 }
 
 pub(super) async fn generate_stream(
@@ -585,515 +546,13 @@ fn transform_chunk_for_provider(
         // Anthropic already emits `{"type":"content_block_delta","delta":{"type":"text_delta","text":"..."}}`.
         // Pass through.
         HttpBedrockProvider::Anthropic | HttpBedrockProvider::Other => Some(decoded.to_string()),
-        HttpBedrockProvider::Nova => transform_nova_chunk_to_anthropic(decoded),
-        HttpBedrockProvider::Llama => transform_llama_chunk_to_anthropic(decoded),
-        HttpBedrockProvider::Mistral => transform_mistral_chunk_to_anthropic(decoded),
-        HttpBedrockProvider::DeepSeek => transform_deepseek_chunk_to_anthropic(decoded),
-        HttpBedrockProvider::Cohere => transform_cohere_chunk_to_anthropic(decoded),
-        HttpBedrockProvider::Ai21Jamba => transform_ai21_jamba_chunk_to_anthropic(decoded),
+        HttpBedrockProvider::Nova => nova::transform_chunk_to_anthropic(decoded),
+        HttpBedrockProvider::Llama => llama::transform_chunk_to_anthropic(decoded),
+        HttpBedrockProvider::Mistral => mistral::transform_chunk_to_anthropic(decoded),
+        HttpBedrockProvider::DeepSeek => deepseek::transform_chunk_to_anthropic(decoded),
+        HttpBedrockProvider::Cohere => cohere::transform_chunk_to_anthropic(decoded),
+        HttpBedrockProvider::Ai21Jamba => ai21_jamba::transform_chunk_to_anthropic(decoded),
     }
-}
-
-/// Llama non-stream response shape (per AWS Bedrock User Guide
-/// `model-parameters-meta.md`):
-/// ```json
-/// { "generation": "...", "prompt_token_count": N, "generation_token_count": M, "stop_reason": "stop" }
-/// ```
-/// We translate it into a single-block Claude payload so the existing
-/// `normalize_claude_response` can fold it into an OpenAI `chat.completion`.
-/// Llama's `stop_reason` values (`stop`, `length`) already align with Claude's
-/// `end_turn` / `max_tokens` after the Claude finish-reason mapping.
-fn llama_response_to_claude_shape(body: Value) -> Value {
-    let text = body
-        .get("generation")
-        .and_then(Value::as_str)
-        .unwrap_or_default()
-        .to_string();
-    let stop_reason = match body.get("stop_reason").and_then(Value::as_str) {
-        Some("length") => "max_tokens".to_string(),
-        Some(other) => other.to_string(),
-        None => "end_turn".to_string(),
-    };
-
-    let mut usage = Map::new();
-    if let Some(input_tokens) = body.get("prompt_token_count").and_then(Value::as_u64) {
-        usage.insert("input_tokens".to_string(), json!(input_tokens));
-    }
-    if let Some(output_tokens) = body.get("generation_token_count").and_then(Value::as_u64) {
-        usage.insert("output_tokens".to_string(), json!(output_tokens));
-    }
-
-    let mut claude_body = Map::new();
-    claude_body.insert(
-        "content".to_string(),
-        Value::Array(vec![json!({ "type": "text", "text": text })]),
-    );
-    claude_body.insert("stop_reason".to_string(), Value::String(stop_reason));
-    if !usage.is_empty() {
-        claude_body.insert("usage".to_string(), Value::Object(usage));
-    }
-    Value::Object(claude_body)
-}
-
-/// Llama stream chunks each carry the next token group in `generation`. The
-/// terminal chunk also carries `stop_reason` so the frontend never sees
-/// `null`. We map every `generation` value (including the empty trailing one)
-/// to an Anthropic `content_block_delta` text frame.
-fn transform_llama_chunk_to_anthropic(decoded: &str) -> Option<String> {
-    let value: Value = serde_json::from_str(decoded).ok()?;
-    let text = value.get("generation").and_then(Value::as_str)?;
-    if text.is_empty() {
-        return None;
-    }
-    Some(
-        json!({
-            "type": "content_block_delta",
-            "index": 0,
-            "delta": { "type": "text_delta", "text": text },
-        })
-        .to_string(),
-    )
-}
-
-/// Mistral on Bedrock exposes two response shapes, switched by model id:
-///
-/// - **Legacy text-completion** (7B, Mixtral, large-2402): the body is
-///   `{ "outputs": [{ "text": "...", "stop_reason": "stop" }] }`.
-/// - **Chat-completion** (large-2407+, small, medium, Pixtral): the body may
-///   be `{ "content": [{ "role": "assistant", "content": [{ "text": "..." }] }] }`
-///   (mistral-large-2407 spec) **or** `{ "choices": [{ "index":0,
-///   "message": { "role":"assistant", "content":"string" }, "stop_reason":"stop" }] }`
-///   (the generic Mistral chat-completion spec).
-///
-/// We probe each shape in turn and emit a Claude-style payload so the existing
-/// `normalize_claude_response` does the OpenAI-shape rewriting.
-fn mistral_response_to_claude_shape(body: Value) -> Value {
-    let text = if let Some(outputs) = body.get("outputs").and_then(Value::as_array) {
-        outputs
-            .iter()
-            .filter_map(|output| output.get("text").and_then(Value::as_str))
-            .collect::<Vec<_>>()
-            .join("")
-    } else if let Some(choices) = body.get("choices").and_then(Value::as_array) {
-        choices
-            .iter()
-            .filter_map(|choice| {
-                choice
-                    .pointer("/message/content")
-                    .and_then(Value::as_str)
-                    .map(str::to_string)
-            })
-            .collect::<Vec<_>>()
-            .join("")
-    } else if let Some(content) = body.get("content").and_then(Value::as_array) {
-        content
-            .iter()
-            .filter_map(|message| {
-                message
-                    .get("content")
-                    .and_then(Value::as_array)
-                    .map(|parts| {
-                        parts
-                            .iter()
-                            .filter_map(|part| {
-                                part.get("text").and_then(Value::as_str).map(str::to_string)
-                            })
-                            .collect::<Vec<_>>()
-                            .join("")
-                    })
-            })
-            .collect::<Vec<_>>()
-            .join("")
-    } else {
-        String::new()
-    };
-
-    let stop_reason = body
-        .pointer("/choices/0/stop_reason")
-        .or_else(|| body.pointer("/outputs/0/stop_reason"))
-        .and_then(Value::as_str)
-        .map(|value| match value {
-            "length" => "max_tokens".to_string(),
-            other => other.to_string(),
-        })
-        .unwrap_or_else(|| "end_turn".to_string());
-
-    let mut usage = Map::new();
-    if let Some(usage_value) = body.get("usage").and_then(Value::as_object) {
-        if let Some(input_tokens) = usage_value
-            .get("prompt_tokens")
-            .or_else(|| usage_value.get("input_tokens"))
-            .and_then(Value::as_u64)
-        {
-            usage.insert("input_tokens".to_string(), json!(input_tokens));
-        }
-        if let Some(output_tokens) = usage_value
-            .get("completion_tokens")
-            .or_else(|| usage_value.get("output_tokens"))
-            .and_then(Value::as_u64)
-        {
-            usage.insert("output_tokens".to_string(), json!(output_tokens));
-        }
-    }
-
-    let mut claude_body = Map::new();
-    claude_body.insert(
-        "content".to_string(),
-        Value::Array(vec![json!({ "type": "text", "text": text })]),
-    );
-    claude_body.insert("stop_reason".to_string(), Value::String(stop_reason));
-    if !usage.is_empty() {
-        claude_body.insert("usage".to_string(), Value::Object(usage));
-    }
-    Value::Object(claude_body)
-}
-
-/// Mistral stream chunks come in three flavours depending on the model:
-///
-/// - Legacy text-completion: `{ "outputs": [{ "text": "...", "stop_reason": null }] }`
-/// - Chat (large-2407 spec):  `{ "content": [{ "text": "..." }] }`
-/// - Chat (chat-completion):  `{ "choices": [{ "delta": { "content": "..." } }] }`
-///
-/// We probe each one and rewrap the extracted text as an Anthropic
-/// `content_block_delta` frame. Chunks without user-visible text (final
-/// `stop_reason` markers, empty `content`, ...) are silently dropped.
-fn transform_mistral_chunk_to_anthropic(decoded: &str) -> Option<String> {
-    let value: Value = serde_json::from_str(decoded).ok()?;
-
-    let text = value
-        .pointer("/choices/0/delta/content")
-        .and_then(Value::as_str)
-        .or_else(|| value.pointer("/content/0/text").and_then(Value::as_str))
-        .or_else(|| value.pointer("/outputs/0/text").and_then(Value::as_str))?;
-
-    if text.is_empty() {
-        return None;
-    }
-
-    Some(
-        json!({
-            "type": "content_block_delta",
-            "index": 0,
-            "delta": { "type": "text_delta", "text": text },
-        })
-        .to_string(),
-    )
-}
-
-/// DeepSeek on Bedrock exposes two response shapes, switched by model id:
-///
-/// - **Text-completion (R1)**: `{ "choices": [{ "text": "...", "stop_reason": "stop"|"length" }] }`.
-/// - **Chat-completion (V3.1+ / V3.2)**: AWS model cards only show the
-///   request side and pretty-print the response via `json.loads(...)`. The
-///   most plausible shape (matching DeepSeek's own OpenAI-compatible API) is
-///   `{ "choices": [{ "message": { "role":"assistant","content":"..." }, "finish_reason":"stop" }] }`,
-///   but Bedrock could also keep the legacy `text` field. We probe both
-///   `choices[0].message.content` and `choices[0].text` defensively.
-///
-/// The result is rewritten into a Claude-shaped payload so the existing
-/// `normalize_claude_response` handles the OpenAI projection.
-fn deepseek_response_to_claude_shape(body: Value) -> Value {
-    let text = body
-        .pointer("/choices/0/message/content")
-        .and_then(Value::as_str)
-        .map(str::to_string)
-        .or_else(|| {
-            body.pointer("/choices/0/text")
-                .and_then(Value::as_str)
-                .map(str::to_string)
-        })
-        .unwrap_or_default();
-
-    let stop_reason = body
-        .pointer("/choices/0/finish_reason")
-        .or_else(|| body.pointer("/choices/0/stop_reason"))
-        .and_then(Value::as_str)
-        .map(|value| match value {
-            "length" => "max_tokens".to_string(),
-            other => other.to_string(),
-        })
-        .unwrap_or_else(|| "end_turn".to_string());
-
-    let mut usage = Map::new();
-    if let Some(usage_value) = body.get("usage").and_then(Value::as_object) {
-        if let Some(input_tokens) = usage_value
-            .get("prompt_tokens")
-            .or_else(|| usage_value.get("input_tokens"))
-            .and_then(Value::as_u64)
-        {
-            usage.insert("input_tokens".to_string(), json!(input_tokens));
-        }
-        if let Some(output_tokens) = usage_value
-            .get("completion_tokens")
-            .or_else(|| usage_value.get("output_tokens"))
-            .and_then(Value::as_u64)
-        {
-            usage.insert("output_tokens".to_string(), json!(output_tokens));
-        }
-    }
-
-    let mut claude_body = Map::new();
-    claude_body.insert(
-        "content".to_string(),
-        Value::Array(vec![json!({ "type": "text", "text": text })]),
-    );
-    claude_body.insert("stop_reason".to_string(), Value::String(stop_reason));
-    if !usage.is_empty() {
-        claude_body.insert("usage".to_string(), Value::Object(usage));
-    }
-    Value::Object(claude_body)
-}
-
-/// DeepSeek stream chunks (per the published R1 schema and the Bedrock model
-/// card examples) come in two flavours:
-///
-/// - Text-completion (R1): `{ "choices": [{ "text": "...", "stop_reason": null|"stop" }] }`
-/// - Chat-completion (V3.1+): `{ "choices": [{ "delta": { "content": "..." } }] }`
-///   (OpenAI-compatible — same as DeepSeek's own /v1/chat/completions stream).
-///
-/// We probe each one and rewrap the extracted text as an Anthropic
-/// `content_block_delta`. Terminal frames without user-visible text (empty
-/// `text` / empty `delta.content`) are silently dropped so the frontend never
-/// re-renders a blank tail.
-fn transform_deepseek_chunk_to_anthropic(decoded: &str) -> Option<String> {
-    let value: Value = serde_json::from_str(decoded).ok()?;
-    let text = value
-        .pointer("/choices/0/delta/content")
-        .and_then(Value::as_str)
-        .or_else(|| value.pointer("/choices/0/text").and_then(Value::as_str))?;
-
-    if text.is_empty() {
-        return None;
-    }
-
-    Some(
-        json!({
-            "type": "content_block_delta",
-            "index": 0,
-            "delta": { "type": "text_delta", "text": text },
-        })
-        .to_string(),
-    )
-}
-
-/// AI21 Jamba non-stream responses (per AI21 chat-completion spec /
-/// Bedrock model card `model-parameters-jamba.md`):
-///
-/// ```json
-/// { "id": "...", "choices": [{ "index": 0,
-///     "message": { "role": "assistant", "content": "..." },
-///     "finish_reason": "stop|length|content_filter" }],
-///   "usage": { "prompt_tokens": N, "completion_tokens": M } }
-/// ```
-///
-/// We lift `choices[0].message.content` into a Claude-shaped `content[].text`
-/// block, project Jamba's `finish_reason` (`length` -> `max_tokens`, `stop` /
-/// `content_filter` -> `end_turn` / `content_filter`), and forward
-/// `usage.prompt_tokens` / `usage.completion_tokens`.
-fn ai21_jamba_response_to_claude_shape(body: Value) -> Value {
-    let text = body
-        .pointer("/choices/0/message/content")
-        .and_then(Value::as_str)
-        .map(str::to_string)
-        .unwrap_or_default();
-
-    let stop_reason = body
-        .pointer("/choices/0/finish_reason")
-        .and_then(Value::as_str)
-        .map(|value| match value {
-            "length" => "max_tokens".to_string(),
-            "stop" => "end_turn".to_string(),
-            other => other.to_string(),
-        })
-        .unwrap_or_else(|| "end_turn".to_string());
-
-    let mut usage = Map::new();
-    if let Some(usage_value) = body.get("usage").and_then(Value::as_object) {
-        if let Some(input_tokens) = usage_value.get("prompt_tokens").and_then(Value::as_u64) {
-            usage.insert("input_tokens".to_string(), json!(input_tokens));
-        }
-        if let Some(output_tokens) = usage_value
-            .get("completion_tokens")
-            .and_then(Value::as_u64)
-        {
-            usage.insert("output_tokens".to_string(), json!(output_tokens));
-        }
-    }
-
-    let mut claude_body = Map::new();
-    claude_body.insert(
-        "content".to_string(),
-        Value::Array(vec![json!({ "type": "text", "text": text })]),
-    );
-    claude_body.insert("stop_reason".to_string(), Value::String(stop_reason));
-    if !usage.is_empty() {
-        claude_body.insert("usage".to_string(), Value::Object(usage));
-    }
-    Value::Object(claude_body)
-}
-
-/// AI21 Jamba stream chunks are OpenAI-shape:
-///
-/// ```json
-/// { "id": "...", "choices": [{
-///     "index": 0,
-///     "delta": { "role": "assistant", "content": "..." },
-///     "finish_reason": null | "stop" | "length"
-/// }] }
-/// ```
-///
-/// The first chunk carries `delta.role = "assistant"` with no content; the
-/// terminal chunk has `delta = {}` (or empty content) and a `usage` summary.
-/// We extract `choices[0].delta.content`, drop empty or sentinel chunks, and
-/// rewrap as an Anthropic `content_block_delta` text frame.
-fn transform_ai21_jamba_chunk_to_anthropic(decoded: &str) -> Option<String> {
-    let value: Value = serde_json::from_str(decoded).ok()?;
-    let text = value
-        .pointer("/choices/0/delta/content")
-        .and_then(Value::as_str)?;
-
-    if text.is_empty() {
-        return None;
-    }
-
-    Some(
-        json!({
-            "type": "content_block_delta",
-            "index": 0,
-            "delta": { "type": "text_delta", "text": text },
-        })
-        .to_string(),
-    )
-}
-
-/// Cohere Command R / R+ non-stream responses (per AWS Bedrock User Guide
-/// `model-parameters-cohere-command-r-plus.md`) look like:
-///
-/// ```json
-/// { "response_id": "...", "text": "...", "finish_reason": "complete|max_tokens|error|...",
-///   "meta": { "billed_units": { "input_tokens": N, "output_tokens": M } } }
-/// ```
-///
-/// We lift `text` straight into a Claude `content[].text` block, project
-/// Cohere's `finish_reason` into Claude vocabulary (`complete` -> `end_turn`,
-/// `max_tokens` stays, `error_*` -> `error`, ...), and forward
-/// `meta.billed_units` as `usage.input_tokens` / `usage.output_tokens`.
-fn cohere_response_to_claude_shape(body: Value) -> Value {
-    let text = body
-        .get("text")
-        .and_then(Value::as_str)
-        .map(str::to_string)
-        .unwrap_or_default();
-
-    let stop_reason = body
-        .get("finish_reason")
-        .and_then(Value::as_str)
-        .map(map_cohere_finish_reason_to_claude)
-        .unwrap_or_else(|| "end_turn".to_string());
-
-    let mut usage = Map::new();
-    if let Some(billed) = body
-        .pointer("/meta/billed_units")
-        .and_then(Value::as_object)
-    {
-        if let Some(input_tokens) = billed.get("input_tokens").and_then(Value::as_u64) {
-            usage.insert("input_tokens".to_string(), json!(input_tokens));
-        }
-        if let Some(output_tokens) = billed.get("output_tokens").and_then(Value::as_u64) {
-            usage.insert("output_tokens".to_string(), json!(output_tokens));
-        }
-    }
-
-    let mut claude_body = Map::new();
-    claude_body.insert(
-        "content".to_string(),
-        Value::Array(vec![json!({ "type": "text", "text": text })]),
-    );
-    claude_body.insert("stop_reason".to_string(), Value::String(stop_reason));
-    if !usage.is_empty() {
-        claude_body.insert("usage".to_string(), Value::Object(usage));
-    }
-    Value::Object(claude_body)
-}
-
-fn map_cohere_finish_reason_to_claude(reason: &str) -> String {
-    match reason {
-        "complete" => "end_turn".to_string(),
-        "max_tokens" => "max_tokens".to_string(),
-        // error_toxic / error_limit / error / user_cancel collapse into the
-        // generic "error" — Anthropic doesn't enumerate a richer vocabulary.
-        other if other.starts_with("error") => "error".to_string(),
-        "user_cancel" => "error".to_string(),
-        other => other.to_string(),
-    }
-}
-
-/// Cohere streams chat events keyed by `event_type` (per
-/// `bedrock-runtime_example_bedrock-runtime_InvokeModelWithResponseStream_CohereCommandR_section`
-/// and the Cohere openapi spec):
-///
-/// - `{ "event_type": "stream-start",   "generation_id": "..." }`
-/// - `{ "event_type": "text-generation","text": "...", "is_finished": false }`
-/// - `{ "event_type": "stream-end",     "finish_reason": "...", "response": {...} }`
-/// - `{ "event_type": "citation-generation" | "tool-calls-*" | ... }`
-///
-/// Only `text-generation` carries user-visible text; everything else is
-/// dropped. We also accept legacy Cohere Command (`generations[0].text`)
-/// chunks defensively so the same path works for `cohere.command-text-v14`
-/// stream output if it ever lands here.
-fn transform_cohere_chunk_to_anthropic(decoded: &str) -> Option<String> {
-    let value: Value = serde_json::from_str(decoded).ok()?;
-    let text = match value.get("event_type").and_then(Value::as_str) {
-        Some("text-generation") => value.get("text").and_then(Value::as_str)?,
-        Some(_) => return None,
-        None => value
-            .pointer("/generations/0/text")
-            .and_then(Value::as_str)?,
-    };
-
-    if text.is_empty() {
-        return None;
-    }
-
-    Some(
-        json!({
-            "type": "content_block_delta",
-            "index": 0,
-            "delta": { "type": "text_delta", "text": text },
-        })
-        .to_string(),
-    )
-}
-
-/// Nova streams Converse-style events. Each EventStream `bytes` chunk decodes
-/// into one of:
-///   - `{ "messageStart": { "role": "assistant" } }`
-///   - `{ "contentBlockStart": { "start": {...}, "contentBlockIndex": 0 } }`
-///   - `{ "contentBlockDelta": { "delta": { "text": "..." }, "contentBlockIndex": 0 } }`
-///   - `{ "contentBlockStop": { "contentBlockIndex": 0 } }`
-///   - `{ "messageStop": { "stopReason": "end_turn" } }`
-///   - `{ "metadata": { "usage": {...} } }`
-///
-/// Only the `contentBlockDelta.delta.text` payload carries user-visible text;
-/// every other event is dropped. The delta is rewrapped as an Anthropic
-/// `content_block_delta` text frame.
-fn transform_nova_chunk_to_anthropic(decoded: &str) -> Option<String> {
-    let value: Value = serde_json::from_str(decoded).ok()?;
-    let delta_text = value
-        .pointer("/contentBlockDelta/delta/text")
-        .and_then(Value::as_str)?;
-    let index = value
-        .pointer("/contentBlockDelta/contentBlockIndex")
-        .and_then(Value::as_u64)
-        .unwrap_or(0);
-    Some(
-        json!({
-            "type": "content_block_delta",
-            "index": index,
-            "delta": { "type": "text_delta", "text": delta_text },
-        })
-        .to_string(),
-    )
 }
 
 #[cfg(test)]
@@ -1104,19 +563,11 @@ mod tests {
     use tokio::sync::mpsc::unbounded_channel;
 
     use super::{
-        HttpBedrockProvider, ai21_jamba_response_to_claude_shape, cohere_response_to_claude_shape,
-        decode_eventstream_payload, deepseek_response_to_claude_shape,
-        derive_control_plane_base, detect_http_provider, drain_eventstream_messages,
-        extract_model_id_from_endpoint, extract_provider, inference_supports_on_demand,
-        llama_response_to_claude_shape, merge_bedrock_models,
-        mistral_response_to_claude_shape, normalize_provider_response,
-        nova_response_to_claude_shape, to_stream_endpoint,
-        transform_ai21_jamba_chunk_to_anthropic, transform_cohere_chunk_to_anthropic,
-        transform_deepseek_chunk_to_anthropic, transform_llama_chunk_to_anthropic,
-        transform_mistral_chunk_to_anthropic, transform_nova_chunk_to_anthropic,
-        validate_invoke_endpoint,
+        HttpBedrockProvider, decode_eventstream_payload, derive_control_plane_base,
+        detect_http_provider, drain_eventstream_messages, extract_model_id_from_endpoint,
+        extract_provider, inference_supports_on_demand, merge_bedrock_models,
+        normalize_provider_response, to_stream_endpoint, validate_invoke_endpoint,
     };
-    use serde_json::Value;
 
     #[test]
     fn validate_invoke_endpoint_accepts_invoke_suffix() {
@@ -1266,356 +717,6 @@ mod tests {
             "anthropic.claude-3-haiku-20240307-v1:0",
         );
         assert_eq!(extract_model_id_from_endpoint("/chat/completions"), "");
-    }
-
-    #[test]
-    fn transform_nova_chunk_to_anthropic_extracts_content_block_delta_text() {
-        let chunk = json!({
-            "contentBlockDelta": {
-                "delta": { "text": "hello world" },
-                "contentBlockIndex": 0,
-            }
-        })
-        .to_string();
-        let rewritten = transform_nova_chunk_to_anthropic(&chunk).expect("delta chunk");
-        let parsed: Value = serde_json::from_str(&rewritten).unwrap();
-        assert_eq!(parsed["type"], "content_block_delta");
-        assert_eq!(parsed["delta"]["text"], "hello world");
-        assert_eq!(parsed["index"], 0);
-    }
-
-    #[test]
-    fn transform_nova_chunk_to_anthropic_drops_non_text_envelopes() {
-        for chunk in [
-            json!({ "messageStart": { "role": "assistant" } }).to_string(),
-            json!({ "messageStop": { "stopReason": "end_turn" } }).to_string(),
-            json!({ "metadata": { "usage": { "inputTokens": 1, "outputTokens": 1 } } })
-                .to_string(),
-        ] {
-            assert!(
-                transform_nova_chunk_to_anthropic(&chunk).is_none(),
-                "non-delta chunk should be silently dropped: {chunk}",
-            );
-        }
-    }
-
-    #[test]
-    fn nova_response_to_claude_shape_lifts_text_and_usage() {
-        let nova_body = json!({
-            "output": {
-                "message": {
-                    "role": "assistant",
-                    "content": [
-                        { "text": "first" },
-                        { "text": "second" }
-                    ]
-                }
-            },
-            "stopReason": "end_turn",
-            "usage": { "inputTokens": 10, "outputTokens": 5, "totalTokens": 15 }
-        });
-
-        let claude_shape = nova_response_to_claude_shape(nova_body);
-        assert_eq!(claude_shape["stop_reason"], "end_turn");
-        let content = claude_shape["content"].as_array().expect("content");
-        assert_eq!(content.len(), 2);
-        assert_eq!(content[0]["type"], "text");
-        assert_eq!(content[0]["text"], "first");
-        assert_eq!(claude_shape["usage"]["input_tokens"], 10);
-        assert_eq!(claude_shape["usage"]["output_tokens"], 5);
-    }
-
-    #[test]
-    fn transform_llama_chunk_to_anthropic_extracts_generation_text() {
-        let chunk = json!({
-            "generation": " world",
-            "prompt_token_count": 10,
-            "generation_token_count": 1,
-            "stop_reason": null,
-        })
-        .to_string();
-        let rewritten = transform_llama_chunk_to_anthropic(&chunk).expect("delta chunk");
-        let parsed: Value = serde_json::from_str(&rewritten).unwrap();
-        assert_eq!(parsed["type"], "content_block_delta");
-        assert_eq!(parsed["delta"]["text"], " world");
-    }
-
-    #[test]
-    fn transform_llama_chunk_to_anthropic_drops_trailing_empty_generation() {
-        let chunk = json!({ "generation": "", "stop_reason": "stop" }).to_string();
-        assert!(transform_llama_chunk_to_anthropic(&chunk).is_none());
-    }
-
-    #[test]
-    fn llama_response_to_claude_shape_lifts_generation_text_and_token_counts() {
-        let body = json!({
-            "generation": "hello world",
-            "prompt_token_count": 12,
-            "generation_token_count": 4,
-            "stop_reason": "stop"
-        });
-        let claude_shape = llama_response_to_claude_shape(body);
-        assert_eq!(claude_shape["stop_reason"], "stop");
-        let content = claude_shape["content"].as_array().expect("content");
-        assert_eq!(content[0]["text"], "hello world");
-        assert_eq!(claude_shape["usage"]["input_tokens"], 12);
-        assert_eq!(claude_shape["usage"]["output_tokens"], 4);
-    }
-
-    #[test]
-    fn llama_response_length_stop_reason_maps_to_claude_max_tokens() {
-        let body = json!({ "generation": "...", "stop_reason": "length" });
-        let claude_shape = llama_response_to_claude_shape(body);
-        assert_eq!(claude_shape["stop_reason"], "max_tokens");
-    }
-
-    #[test]
-    fn transform_mistral_chunk_to_anthropic_handles_all_three_shapes() {
-        let legacy = json!({ "outputs": [{ "text": " hi", "stop_reason": null }] }).to_string();
-        let chat_2407 = json!({ "content": [{ "text": " from" }] }).to_string();
-        let chat_openai = json!({
-            "choices": [{ "index": 0, "delta": { "content": " mistral" } }],
-        })
-        .to_string();
-
-        for (chunk, expected) in [
-            (legacy, " hi"),
-            (chat_2407, " from"),
-            (chat_openai, " mistral"),
-        ] {
-            let rewritten = transform_mistral_chunk_to_anthropic(&chunk)
-                .unwrap_or_else(|| panic!("expected delta from chunk: {chunk}"));
-            let parsed: Value = serde_json::from_str(&rewritten).unwrap();
-            assert_eq!(parsed["type"], "content_block_delta");
-            assert_eq!(parsed["delta"]["text"], expected);
-        }
-    }
-
-    #[test]
-    fn transform_mistral_chunk_to_anthropic_drops_terminal_stop_reason_chunks() {
-        let chunk = json!({ "outputs": [{ "text": "", "stop_reason": "stop" }] }).to_string();
-        assert!(transform_mistral_chunk_to_anthropic(&chunk).is_none());
-    }
-
-    #[test]
-    fn mistral_response_to_claude_shape_picks_up_text_from_each_response_dialect() {
-        let legacy = json!({
-            "outputs": [{ "text": "legacy text", "stop_reason": "stop" }]
-        });
-        let chat_2407 = json!({
-            "content": [{
-                "role": "assistant",
-                "content": [{ "text": "chat 2407 text" }]
-            }]
-        });
-        let chat_openai = json!({
-            "choices": [{
-                "index": 0,
-                "message": { "role": "assistant", "content": "chat openai text" },
-                "stop_reason": "length"
-            }]
-        });
-
-        let legacy_shape = mistral_response_to_claude_shape(legacy);
-        assert_eq!(legacy_shape["content"][0]["text"], "legacy text");
-        assert_eq!(legacy_shape["stop_reason"], "stop");
-
-        let chat_2407_shape = mistral_response_to_claude_shape(chat_2407);
-        assert_eq!(chat_2407_shape["content"][0]["text"], "chat 2407 text");
-
-        let chat_openai_shape = mistral_response_to_claude_shape(chat_openai);
-        assert_eq!(chat_openai_shape["content"][0]["text"], "chat openai text");
-        assert_eq!(
-            chat_openai_shape["stop_reason"], "max_tokens",
-            "length must map to Claude max_tokens",
-        );
-    }
-
-    #[test]
-    fn transform_deepseek_chunk_to_anthropic_handles_text_and_delta_shapes() {
-        let r1 = json!({
-            "choices": [{ "text": " hello", "stop_reason": null }],
-        })
-        .to_string();
-        let v3 = json!({
-            "choices": [{ "delta": { "content": " world" } }],
-        })
-        .to_string();
-
-        for (chunk, expected) in [(r1, " hello"), (v3, " world")] {
-            let rewritten = transform_deepseek_chunk_to_anthropic(&chunk)
-                .unwrap_or_else(|| panic!("expected delta from chunk: {chunk}"));
-            let parsed: Value = serde_json::from_str(&rewritten).unwrap();
-            assert_eq!(parsed["type"], "content_block_delta");
-            assert_eq!(parsed["delta"]["text"], expected);
-        }
-    }
-
-    #[test]
-    fn transform_deepseek_chunk_to_anthropic_drops_terminal_empty_text() {
-        let chunk = json!({
-            "choices": [{ "text": "", "stop_reason": "stop" }],
-        })
-        .to_string();
-        assert!(transform_deepseek_chunk_to_anthropic(&chunk).is_none());
-    }
-
-    #[test]
-    fn deepseek_response_to_claude_shape_lifts_text_and_token_counts() {
-        let r1 = json!({
-            "choices": [{ "text": "r1 reply", "stop_reason": "length" }],
-            "usage": { "prompt_tokens": 10, "completion_tokens": 5 },
-        });
-        let r1_shape = deepseek_response_to_claude_shape(r1);
-        assert_eq!(r1_shape["content"][0]["text"], "r1 reply");
-        assert_eq!(r1_shape["stop_reason"], "max_tokens");
-        assert_eq!(r1_shape["usage"]["input_tokens"], 10);
-        assert_eq!(r1_shape["usage"]["output_tokens"], 5);
-
-        let v3 = json!({
-            "choices": [{
-                "message": { "role": "assistant", "content": "v3 reply" },
-                "finish_reason": "stop"
-            }]
-        });
-        let v3_shape = deepseek_response_to_claude_shape(v3);
-        assert_eq!(v3_shape["content"][0]["text"], "v3 reply");
-        assert_eq!(v3_shape["stop_reason"], "stop");
-    }
-
-    #[test]
-    fn transform_cohere_chunk_to_anthropic_only_keeps_text_generation_events() {
-        let text_gen = json!({
-            "event_type": "text-generation",
-            "text": " hi",
-            "is_finished": false,
-        })
-        .to_string();
-        let rewritten = transform_cohere_chunk_to_anthropic(&text_gen)
-            .expect("text-generation event must surface a delta");
-        let parsed: Value = serde_json::from_str(&rewritten).unwrap();
-        assert_eq!(parsed["type"], "content_block_delta");
-        assert_eq!(parsed["delta"]["text"], " hi");
-
-        let stream_start = json!({
-            "event_type": "stream-start",
-            "generation_id": "abc",
-        })
-        .to_string();
-        assert!(transform_cohere_chunk_to_anthropic(&stream_start).is_none());
-
-        let stream_end = json!({
-            "event_type": "stream-end",
-            "finish_reason": "complete",
-            "response": {},
-        })
-        .to_string();
-        assert!(transform_cohere_chunk_to_anthropic(&stream_end).is_none());
-
-        // Legacy Cohere Command (generations[0].text) — kept as a defensive
-        // fallback for cohere.command-text-v14 / command-light-text-v14 if
-        // they ever stream through here.
-        let legacy = json!({
-            "generations": [{ "text": " legacy" }]
-        })
-        .to_string();
-        let legacy_delta = transform_cohere_chunk_to_anthropic(&legacy)
-            .expect("legacy chunks must surface a delta");
-        let parsed: Value = serde_json::from_str(&legacy_delta).unwrap();
-        assert_eq!(parsed["delta"]["text"], " legacy");
-    }
-
-    #[test]
-    fn cohere_response_to_claude_shape_maps_finish_reason_and_billed_units() {
-        let body = json!({
-            "response_id": "abc",
-            "text": "cohere reply",
-            "finish_reason": "max_tokens",
-            "meta": {
-                "billed_units": { "input_tokens": 42, "output_tokens": 7 }
-            }
-        });
-
-        let shape = cohere_response_to_claude_shape(body);
-        assert_eq!(shape["content"][0]["text"], "cohere reply");
-        assert_eq!(shape["stop_reason"], "max_tokens");
-        assert_eq!(shape["usage"]["input_tokens"], 42);
-        assert_eq!(shape["usage"]["output_tokens"], 7);
-
-        let complete = json!({ "text": "done", "finish_reason": "complete" });
-        assert_eq!(
-            cohere_response_to_claude_shape(complete)["stop_reason"],
-            "end_turn",
-            "complete must map to Claude end_turn",
-        );
-
-        let toxic = json!({ "text": "", "finish_reason": "error_toxic" });
-        assert_eq!(
-            cohere_response_to_claude_shape(toxic)["stop_reason"],
-            "error",
-            "any error_* must collapse into Claude error",
-        );
-    }
-
-    #[test]
-    fn transform_ai21_jamba_chunk_to_anthropic_extracts_delta_content() {
-        let chunk = json!({
-            "id": "abc",
-            "choices": [{
-                "index": 0,
-                "delta": { "content": " jamba" },
-                "finish_reason": null
-            }]
-        })
-        .to_string();
-        let rewritten = transform_ai21_jamba_chunk_to_anthropic(&chunk)
-            .expect("jamba delta chunk must surface text");
-        let parsed: Value = serde_json::from_str(&rewritten).unwrap();
-        assert_eq!(parsed["type"], "content_block_delta");
-        assert_eq!(parsed["delta"]["text"], " jamba");
-    }
-
-    #[test]
-    fn transform_ai21_jamba_chunk_to_anthropic_drops_role_only_and_terminal_frames() {
-        let role_only = json!({
-            "choices": [{ "delta": { "role": "assistant" } }],
-        })
-        .to_string();
-        assert!(transform_ai21_jamba_chunk_to_anthropic(&role_only).is_none());
-
-        let terminal = json!({
-            "choices": [{ "delta": {}, "finish_reason": "stop" }],
-            "usage": { "prompt_tokens": 1, "completion_tokens": 1 }
-        })
-        .to_string();
-        assert!(transform_ai21_jamba_chunk_to_anthropic(&terminal).is_none());
-    }
-
-    #[test]
-    fn ai21_jamba_response_to_claude_shape_maps_finish_reason_and_usage() {
-        let body = json!({
-            "id": "abc",
-            "choices": [{
-                "index": 0,
-                "message": { "role": "assistant", "content": "jamba reply" },
-                "finish_reason": "length"
-            }],
-            "usage": { "prompt_tokens": 12, "completion_tokens": 4 }
-        });
-        let shape = ai21_jamba_response_to_claude_shape(body);
-        assert_eq!(shape["content"][0]["text"], "jamba reply");
-        assert_eq!(shape["stop_reason"], "max_tokens");
-        assert_eq!(shape["usage"]["input_tokens"], 12);
-        assert_eq!(shape["usage"]["output_tokens"], 4);
-
-        let stop = json!({
-            "choices": [{ "message": { "content": "ok" }, "finish_reason": "stop" }]
-        });
-        assert_eq!(
-            ai21_jamba_response_to_claude_shape(stop)["stop_reason"],
-            "end_turn",
-            "stop must map to Claude end_turn",
-        );
     }
 
     #[test]
